pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = "REST API для управління контактами"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.0"}
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
faker = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"